Fix bug where starting the parser would sometimes not realize pid file was written
Hard to test, b/c it's a little nondeterministic, and the tests have never hit it,
AFAIK. When running from Atom, though, it would always hit this on the first run.
The cause was that it would take the current time, and then start the server, which
would launch and write the portfile quickly enough that there was no perceptible difference
between the pre-launch time and the portfile's mtime.

The solution is to allow them to be the same time (`<=` rather than `<`), but
even that wasn't quite enough. Sometimes Ruby's time would be subsecond more than
the modified time. So, convert to int in order to use a precision of seconds, and
then subtract 1 second, to reduce the probability of precision issues causing
infinite loops. There is some probability that it incorrectly finds a port from
a dead process, but it's low enough, and in practice, not what happens, whereas
this issue with starting the parser happens frequently (eg every 30 min or w/e).
So better to optimize this common case than that unlikely case.

Again, can't think of a good way to test that, so it's just going to be untested
behaviour (I don't want unit tests on the parser's implementation, b/c I could
very conceiveably swap it out for a different parser, or a parser of my own writing).